finalised buttons on register, issue on search
in search pane im having issues connecting buttons to controller due to the methods being of the type void and not returning or not returning a entity as far as i can understand.  changed other classes to make interface work. missing one method to show all results for one student in controller